backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
requests>=2.31.0
argon2-cffi>=23.1.0
